exla/models/clip/_implementations/requirements/requirements_jetson.txt
pillow
transformers
psutil
numpy==1.26.4
https://developer.download.nvidia.com/compute/redist/jp/v61/pytorch/torch-2.5.0a0+872d972e41.nv24.08.17622132-cp310-cp310-linux_aarch64.whl; python_version == "3.10"